python/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=1.10.0
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
mitmproxy>=10.0.0
requests
python-dotenv